修复FP4 decode索引类型错误：uint8转long

+++ apt_model/optimization/gpu_flash_optimization.py
@@ -106,8 +106,8 @@ class FP4Codec:
         # 展平
         indices = indices.flatten(-2)  # [B, N_padded]
 
-        # 查表 - flatten indices for 1D table lookup, then reshape
-        values = table[indices.flatten()].reshape(indices.shape)
+        # 查表 - convert to long for indexing, then lookup and reshape
+        values = table[indices.flatten().long()].reshape(indices.shape)
         
         # 去掉padding
         values = values[..., :original_size]
